+++ CHANGELOG.md
@@ -21,6 +21,9 @@ Changes are grouped as follows
 
 ## [0.7.2] - 2019-11-15
 
+## Added
+- Provide config file path with env var COGNITE_CONFIG_FILE
+
 ### Changed
 - Use version for tag on images published to docker hub
 - Replicate datapoints with a consistent time series order